updated: toast and toast error message, theme and helpers
