+++ .gitignore
@@ -16,6 +16,9 @@ migrate_working_dir/
 *.iws
 .idea/
 
+# VS code
+.vscode/
+
 # The .vscode folder contains launch configuration and tasks you configure in
 # VS Code which you may wish to be included in version control, so this line
 # is commented out by default.
