+++ lib/env.dart
@@ -3,7 +3,8 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
-import 'vaahextendflutter/helpers/console.dart';
+import 'theme.dart';
+import 'vaahextendflutter/helpers/console_log_helper.dart';
 
 // After changing any const you will need to restart the app (Hot-reload won't work).
 
@@ -11,19 +12,19 @@ import 'vaahextendflutter/helpers/console.dart';
 const String version = '1.0.0'; // version format 1.0.0 (major.minor.patch)
 const String build = '2022100901'; // build no format 'YYYYMMDDNUMBER'
 
-EnvironmentConfig defaultConfig = const EnvironmentConfig(
+EnvironmentConfig defaultConfig = EnvironmentConfig(
   envType: 'default',
   version: version,
   build: build,
-  baseUrl: '',
-  apiBaseUrl: 'https://apivoid.herokuapp.com',
+  backendUrl: '',
+  apiUrl: 'https://apivoid.herokuapp.com',
   timeoutLimit: 60 * 1000, // 60 seconds
-  analyticsId: '',
+  firebaseId: '',
   enableConsoleLogs: true,
   enableLocalLogs: true,
   enableApiLogs: true,
   showEnvAndVersionTag: true,
-  envAndVersionTagColor: dangerColor,
+  envAndVersionTagColor: AppTheme.blackColor.withOpacity(0.7),
 );
 
 // To add new configuration add new key, value pair in envConfigs
@@ -75,9 +76,10 @@ class EnvironmentConfig {
   final String envType;
   final String version;
   final String build;
-  final String baseUrl;
-  final String apiBaseUrl;
-  final String analyticsId;
+  final String backendUrl;
+  final String apiUrl;
+  final String firebaseId;
+  final int timeoutLimit;
   final bool enableConsoleLogs;
   final bool enableLocalLogs;
   final bool enableApiLogs;
@@ -88,9 +90,10 @@ class EnvironmentConfig {
     required this.envType,
     required this.version,
     required this.build,
-    required this.baseUrl,
-    required this.apiBaseUrl,
-    required this.analyticsId,
+    required this.backendUrl,
+    required this.apiUrl,
+    required this.firebaseId,
+    required this.timeoutLimit,
     required this.enableConsoleLogs,
     required this.enableLocalLogs,
     required this.enableApiLogs,
@@ -102,9 +105,10 @@ class EnvironmentConfig {
     String? envType,
     String? version,
     String? build,
-    String? baseUrl,
-    String? apiBaseUrl,
-    String? analyticsId,
+    String? backendUrl,
+    String? apiUrl,
+    String? firebaseId,
+    int? timeoutLimit,
     bool? enableConsoleLogs,
     bool? enableLocalLogs,
     bool? enableApiLogs,
@@ -115,9 +119,10 @@ class EnvironmentConfig {
       envType: envType ?? this.envType,
       version: version ?? this.version,
       build: build ?? this.build,
-      baseUrl: baseUrl ?? this.baseUrl,
-      apiBaseUrl: apiBaseUrl ?? this.apiBaseUrl,
-      analyticsId: analyticsId ?? this.analyticsId,
+      backendUrl: backendUrl ?? this.backendUrl,
+      apiUrl: apiUrl ?? this.apiUrl,
+      firebaseId: firebaseId ?? this.firebaseId,
+      timeoutLimit: timeoutLimit ?? this.timeoutLimit,
       enableConsoleLogs: enableConsoleLogs ?? this.enableConsoleLogs,
       enableLocalLogs: enableLocalLogs ?? this.enableLocalLogs,
       enableApiLogs: enableApiLogs ?? this.enableApiLogs,
@@ -127,117 +132,3 @@ class EnvironmentConfig {
     );
   }
 }
-
-const MaterialColor primaryColor = MaterialColor(
-  0xFF3366FF,
-  <int, Color>{
-    50: Color(0xFFD6E4FF),
-    100: Color(0xFFD6E4FF),
-    200: Color(0xFFADC8FF),
-    300: Color(0xFF84A9FF),
-    400: Color(0xFF6690FF),
-    500: Color(0xFF3366FF),
-    600: Color(0xFF254EDB),
-    700: Color(0xFF1939B7),
-    800: Color(0xFF102693),
-    900: Color(0xFF091A7A),
-  },
-);
-
-const MaterialColor successColor = MaterialColor(
-  0xFF4FB52D,
-  <int, Color>{
-    50: Color(0xFFE9FBD5),
-    100: Color(0xFFE9FBD5),
-    200: Color(0xFFCFF7AD),
-    300: Color(0xFFA8E87F),
-    400: Color(0xFF81D25B),
-    500: Color(0xFF4FB52D),
-    600: Color(0xFF369B20),
-    700: Color(0xFF228216),
-    800: Color(0xFF11680E),
-    900: Color(0xFF08560B),
-  },
-);
-
-const MaterialColor infoColor = MaterialColor(
-  0xFF4CA8FF,
-  <int, Color>{
-    50: Color(0xFFDBF4FF),
-    100: Color(0xFFDBF4FF),
-    200: Color(0xFFB7E7FF),
-    300: Color(0xFF93D5FF),
-    400: Color(0xFF78C4FF),
-    500: Color(0xFF4CA8FF),
-    600: Color(0xFF3783DB),
-    700: Color(0xFF2662B7),
-    800: Color(0xFF184493),
-    900: Color(0xFF0E2F7A),
-  },
-);
-
-const MaterialColor warningColor = MaterialColor(
-  0xFFFFBF00,
-  <int, Color>{
-    50: Color(0xFFFFF7CC),
-    100: Color(0xFFFFF7CC),
-    200: Color(0xFFFFED99),
-    300: Color(0xFFFFE066),
-    400: Color(0xFFFFD33F),
-    500: Color(0xFFFFBF00),
-    600: Color(0xFFDB9E00),
-    700: Color(0xFFB77F00),
-    800: Color(0xFF936300),
-    900: Color(0xFF7A4E00),
-  },
-);
-
-const MaterialColor dangerColor = MaterialColor(
-  0xFFFF382D,
-  <int, Color>{
-    50: Color(0xFFFFE5D5),
-    100: Color(0xFFFFE5D5),
-    200: Color(0xFFFFC4AB),
-    300: Color(0xFFFF9C81),
-    400: Color(0xFFFF7661),
-    500: Color(0xFFFF382D),
-    600: Color(0xFFDB2026),
-    700: Color(0xFFB71629),
-    800: Color(0xFF930E28),
-    900: Color(0xFF7A0828),
-  },
-);
-
-
-
-const MaterialColor whiteColor = MaterialColor(
-  0xFFFFFFFF,
-  <int, Color>{
-    50: Color(0xFFFFFFFF),
-    100: Color(0xFFFFFFFF),
-    200: Color(0xFFFFFFFF),
-    300: Color(0xFFFFFFFF),
-    400: Color(0xFFFFFFFF),
-    500: Color(0xFFFFFFFF),
-    600: Color(0xFFFFFFFF),
-    700: Color(0xFFFFFFFF),
-    800: Color(0xFFFFFFFF),
-    900: Color(0xFFFFFFFF),
-  },
-);
-
-const MaterialColor blackColor = MaterialColor(
-  0xFF000000,
-  <int, Color>{
-    50: Color(0xFFF2F2F2),
-    100: Color(0xFFF2F2F2),
-    200: Color(0xFFE5E5E5),
-    300: Color(0xFFB2B2B2),
-    400: Color(0xFF666666),
-    500: Color(0xFF000000),
-    600: Color(0xFF000000),
-    700: Color(0xFF000000),
-    800: Color(0xFF000000),
-    900: Color(0xFF000000),
-  },
-);
