+++ lib/routes/routes.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+import '../view/pages/home/home.dart';
+
+Route<dynamic>? onGenerateRoute(RouteSettings settings) {
+  if (settings.name == '/') {
+    return TeamHomePage.route();
+  }
+  return null;
+}
