+++ lib/vaahextendflutter/tag/tag_panel.dart
@@ -11,8 +11,8 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
 import '../../env.dart';
-import '../helpers/constants.dart';
-import '../helpers/styles.dart';
+import '../helpers/constant_helpers.dart';
+import '../helpers/styles_helper.dart';
 
 const double constHandleWidth = 180.0; // tag handle width
 const double constHandleHeight = 28.0; // tag handle height
@@ -160,11 +160,11 @@ class TagPanelHostState extends State<TagPanelHost>
                                       padding: EdgeInsets.only(
                                         top:
                                             MediaQuery.of(context).padding.top +
-                                                kDeafaultPadding +
+                                                deafaultPadding +
                                                 _handleHeight,
-                                        bottom: kDeafaultPadding,
-                                        left: kDeafaultPadding,
-                                        right: kDeafaultPadding,
+                                        bottom: deafaultPadding,
+                                        left: deafaultPadding,
+                                        right: deafaultPadding,
                                       ),
                                       children: [
                                         SelectableText(
