+++ test/widget_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:team/main.dart';
+import 'package:team/app/app.dart';
 
 void main() {
   testWidgets('test', (WidgetTester tester) async {
